feat: improve default console logger
