+++ APIMatic.Core/Utilities/Logger/SdkLogger.cs
@@ -56,7 +56,7 @@ namespace APIMatic.Core.Utilities.Logger
             var localLogLevel = _logLevel.GetValueOrDefault(LogLevel.Information);
             var contentTypeHeader = response.Headers.GetContentType();
             var contentLengthHeader = response.Headers.GetContentLength();
-            _logger.Log(localLogLevel, "Response {HttpStatusCode} {Length} {ContentType}",
+            _logger.Log(localLogLevel, "Response {HttpStatusCode} {ContentType} ContentLength: {ContentLength}",
                 response.StatusCode,
                 contentLengthHeader,
                 contentTypeHeader);
